% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{aeo_ex}
\alias{aeo_ex}
\title{Emissions by fuel from the 2019 Annual Energy Outlook}
\format{
A dataframe with y rows and x columns
}
\usage{
aeo_ex
}
\description{
Emissions in MMT CO2-Eq. values, for the reference and all side case scenarios from the Energy Information Administration's 2019 AEO. This is used as example data for the envelope geom because the multiple scenarios can be summarized as an envelope.
}
\keyword{datasets}
