% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area-bar-plot.R
\name{area_bar_plot}
\alias{area_bar_plot}
\title{Draw a bar plot with variable widths}
\usage{
area_bar_plot(data, xname, yname, fillname, gap_fixed = 1)
}
\arguments{
\item{data}{to be plotted}

\item{xname}{variable name for x axis (and widths)}

\item{yname}{variable name for y axis (and heights)}

\item{fillname}{variable name for fill colors}

\item{gap_fixed}{a fixed gap width (in units of x)

The standard ggplot2::geom_bar/geom_col has a width argument, but this argument does not affect the placement of the bars, so they quickly overlap. In addition, no x axis is displayed.}
}
\description{
Draw a bar plot with variable widths
}
