% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{ei_ex}
\alias{ei_ex}
\title{Energy-intensity and value-of-shipments by NAICS codes}
\format{A dataframe with y rows and x columns}
\usage{
ei_ex
}
\description{
A dataset containing ...
}
\keyword{datasets}
