% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope-plot.R
\name{envelope_fig}
\alias{envelope_fig}
\title{envelope-plot}
\usage{
envelope_fig(data, xname, yname, groupname, facetname)
}
\arguments{
\item{data}{to be plotted}

\item{xname}{name of x aesthetic variable}

\item{yname}{name of y aesthetic variable}

\item{groupname}{name of variable to define groups}

\item{facetname}{name of facet column variable}
}
\description{
envelope-plot
}
