% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_data.R
\name{interpolate_data}
\alias{interpolate_data}
\title{Interpolation in a dataframe}
\usage{
interpolate_data(.data, x_var, y_var, result_var, interp_var, by_vars)
}
\arguments{
\item{y_var}{string or symbol of the value variable being interpolated}

\item{by_vars}{character vector of variables that define grouping}

\item{data}{within which to interpolate}
}
\description{
Plotting envelopes or computing min/max/percentile statistics over a time series requires that the same year intervals be used in each series. This function is intended to fill in intermediate values through interpolation, in the context of a completed data frame, where NA valeus represent values to be filled.
}
\details{
This function is an awkward fit for this package, but I didn't know where to put it!

The additional variable specifications, result_var and interp_var, can be useful for debugging because they provide the constituent detailed data.
}
